fix: remove camel-case
NOTE: this could cause some replaying tests to fail as the file name has changed.

I don't want to make it a breaking change because the actual API and behavior didn't.

This only apply to file based recoding (which is the only thing supported), and it is better to fix it now than later.