fix: support promise to be invoked later.
Meaning, there can be other action performed (e.g. invoking callbacks) between `return` and `promise/resolve or reject`.

BREAKING CHANGE:
The auto-completing feature is removed.

You should ensure all code has been executed before calling spec.satisfy().

This is more inlined with the actual behavior and better support multiple invocation of the spec.

It is more useful to spec on the gateway towards the boundary for multiple calls, including setup (the Given part of BDD).

So the auto-complete feature is actually not really useful and gives false impression to the test.