feat: rewrite
This is an internal release.
It is released to support building other plugins in order to improve the final API for 5.0 release.
There are some unnecessary `console.log()` messages.
They are there just for development.
Will be removed in the final release. 

Exposed some expectation helpers.

BREAKING CHANGE:

Updated plugin system to use `komondor-plugin`.
The implementation of plugins is vastly simplified.
Some adjustments still need to be made.

The underlying data structure has changed.
It better tracks which return is made by which instance and call.
All spec records need to be re-generated.