feat: add mockto.teardown()
When test fails or when `spec.done()` is not called.

The timeTracker handle is not cleared.
So if the said test is close to the end of the test file,
test runner like jest would emit a warning saying handle not cleared,
and the timeout warning is not printed.

mockto.teardown() will emit the timeout warning and clean up properly.